Revert "fix win streak problem"
This reverts commit 50e40b67757c015a385d9e5cd7811b28301ceda6.

+++ src/main/java/me/fallenbreath/tcuhc/util/PlayerMatchMakingData.java
@@ -24,7 +24,7 @@ public class PlayerMatchMakingData {
         return UUID;
     }
 
-    public void setHisPP(double hisPP) {
+    public void setHisPP(int hisPP) {
         this.hisPP = hisPP;
     }
